Refactor de Store
